feat: update has button story
